add check Token

+++ HelloGitHub/Login/Model/Token.swift
@@ -9,9 +9,9 @@ import Foundation
 
 struct TokenResponse: Codable {
     var access_token: String?
-    var expires_in: Int?
+    var expires_in: Double?
     var refresh_token: String?
-    var refresh_token_expires_in: Int?
+    var refresh_token_expires_in: Double?
     var scope: String?
     var token_type: String?
     var isValid: Bool
